Add enums to represent game object types.
The easiest way to do this was to have the map continue to use the
serialized string -> int mapping and convert to object enum -> hex on
the fly. I could go further with that and (de-)serialize closer to
reading/writing to disk, but I don't think performance will matter.